+++ circuits/src/parsing/scanner/automaton.rs
@@ -36,13 +36,7 @@ use std::{collections::hash_map::Entry, fmt::Debug, hash::Hash, iter::once};
 
 use rustc_hash::{FxBuildHasher, FxHashMap, FxHashSet};
 
-/// Maximal size of the alphabet of an automaton/regex, since input characters
-/// are represented by `AssignedByte`. The parser (`automaton_chip::parse`) is
-/// using this information to store automaton final states in the transition
-/// table, by encoding them as impossible transitions starting from the said
-/// state, and labelled with letter `ALPHABET_MAX_SIZE`. This bound is also
-/// needed to represent letters as u8.
-pub const ALPHABET_MAX_SIZE: usize = 256;
+use crate::parsing::scanner::ALPHABET_MAX_SIZE;
 
 /// A letter from the automaton alphabet. Includes output markers.
 #[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]